% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_lppd.R
\name{compute_lppd}
\alias{compute_lppd}
\title{Compute log pointwise predictive density}
\usage{
compute_lppd(MSOcc_mod)
}
\arguments{
\item{MSOcc_mod}{output from \code{\link{MSOcc_mod}}}
}
\value{
numeric value that is the lppd
}
\description{
This function computes the log pointwise predictive density (lppd)
 described in Gelman et al. (2013) for multi-scale occupancy models.
}
