% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fung}
\alias{fung}
\title{Fungal Pathogen Data}
\format{A data frame with 80 rows and 8 variables:
\describe{
  \item{site}{location sampled}
  \item{sample}{sample number from site}
  \item{pcr1}{detection (1) or non-dection (0) from first PCR replicate}
  \item{pcr2}{detection (1) or non-dection (0) from second PCR replicate}
  \item{loadT1}{average fungal pathogen load of frogs that tested positive for Bd and were captured during 1st survey}
  \item{loadT2}{average fungal pathogen load of frogs that tested positive for Bd and were captured during 2nd survey}
  \item{frogs}{index of frog density (catch per unit effort)}
  \item{Bd}{prevalence of Bd among frogs captured during 1st or 2nd surveys}
  }}
\source{
Schmidt BR, Kery M, Ursenbacher S, Hyman OJ, and Collins JP (2013)
  Site occupancy models in the analysis of environmental DNA presence/absence
  surveys: a case study of an emerging amphibian pathogen. Methods in Ecology
  and Evolution 4: 646-653.
}
\usage{
fung
}
\description{
A data set containing detection of a fungal pathogen's DNA measured at each
of 20 ponds located in Coconino National Forest, Arizona, USA. At each site,
water samples were collected to extract the DNA of Batrachochytrium
dendrobatidis (Bd), a fungal pathogen of various species of amphibians.
Environmental and ecological variables were also measured at each of the 20
ponds, and are included in this dataset.
}
\keyword{datasets}
