% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msocc.R
\name{msocc}
\alias{msocc}
\title{msocc: Fit and analyze computationally efficient multi-scale occupancy models}
\description{
This package provides functions for fitting and analyzing multiscale occupancy models.
}
\seealso{
Useful links:
\itemize{
\item \href{https://pubs.er.usgs.gov/publication/70194000}{Dorazio and Erickson (2017) eDNA occupancy paper}
\item \href{https://arxiv.org/pdf/1205.0310.pdf}{Polson et al. (2012) polya-gamma paper}
\item \href{https://link.springer.com/article/10.1007/s11222-013-9416-2}{Gelman et al. (2013) information criterion paper}
}
}
\author{
\strong{Maintainer:} Christian Stratton \email{christianstratton@montana.edu} \cr \cr
Authors:
\itemize{
\item Dr. Andrew Hoegh
\item Dr. Adam Sepulveda
\item Dr. Katharine Banner
}
}
