% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msocc_mod.R
\name{msocc_mod}
\alias{msocc_mod}
\title{Fit multi-scale occupancy model.}
\usage{
msocc_mod(wide_data, site = list(model = ~1, cov_tbl),
  sample = list(model = ~1, cov_tbl), rep = list(model = ~1, cov_tbl),
  priors = list(site = list(mu0 = 0, Sigma0 = 9), sample = list(mu0 = 0,
  Sigma0 = 9), rep = list(mu0 = 0, Sigma0 = 9), a0 = 1, b0 = 1),
  num.mcmc = 1000, progress = T, print = NULL, seed = NULL,
  beta_bin = T)
}
\arguments{
\item{wide_data}{object of class \code{data.frame} containing site, sample,
and PCR replicates in wide format. Column names should be \code{site},
\code{sample}, \code{PCR1}, \code{PCR2}, ... and contain no other columns.}

\item{site}{object of class \code{list} containing the following elements: \cr
\itemize{
\item \code{model} formula describing the within site model. \cr
\item \code{cov_tbl} object of class \code{data.frame} containing site specific
covariates; \code{cov_tbl} should have exactly one row for each site and have a column named 'site'.
}}

\item{sample}{object of class \code{list} containing the following elements: \cr
\itemize{
\item \code{model} formula describing the within sample model. \cr
\item \code{cov_tbl} object of class \code{data.frame} containing sample specific
covariates; \code{cov_tbl} should have exactly one row for each sample and
have columns named 'site' and 'sample'.
}}

\item{rep}{object of class \code{list} containing the following elements: \cr
\itemize{
\item \code{model} formula describing the within replicate model. \cr
\item \code{cov_tbl} object of class \code{data.frame} containing replicate specific
covariates; \code{cov_tbl} should have exactly one row for each sample if
replicate specific covariates are aggregated at the sample level and contain
columns named 'site' and 'sample'. Otherwise, \code{cov_tbl} should have
exactly one row for each replicate and contain columns name 'site',
'sample', and 'rep'.
}}

\item{priors}{object of class \code{list} containing the following elements: \cr
\itemize{
\item \code{site} \verb{ } object of class \code{list} containing the
following elements: \cr
\itemize{
\item \code{mu0} prior mean for site-level regression coefficients \cr
\item \code{Sigma0} prior covariance matrix for site-level regression
coefficients
}
\item \code{sample} \verb{ } object of class \code{list} containing the
following elements: \cr
\itemize{
\item \code{mu0} prior mean for sample-level regression coefficients \cr
\item \code{Sigma0} prior covariance matrix for sample-level regression
coefficients
}
\item \code{rep} \verb{ } object of class \code{list} containing the
following elements: \cr
\itemize{
\item \code{mu0} prior mean for replicate-level regression coefficients \cr
\item \code{Sigma0} prior covariance matrix for replicate-level regression
coefficients
}
\item \code{a0,b0} numeric shape parameters for beta prior on probability
parameters if sampled directly (see \code{beta_bin} argument)
}}

\item{num.mcmc}{number of MCMC samples}

\item{progress}{should sampling progress be printed?}

\item{print}{interval for printing; defaults to 5 percent of \code{num.mcmc}
of left \code{NULL}}

\item{seed}{optional seed for reproducible samples}

\item{beta_bin}{optional; should a beta-binomial sampler be used when
possible? This option is considerably faster.}
}
\value{
object of class \code{list} containing the following elements: \cr
\itemize{
\item \code{beta} an object of class \code{matrix} of samples from the joint
posterior distribution of the regression coefficients at the site level
\item \code{psi} an object of class \code{numeric} of samples from the joint
posterior distribution of the site-level presence probability, psi. Note
that this is only returned if \code{beta_bin} is \code{TRUE} and
\code{site$model = ~ 1}
\item \code{alpha} an object of class \code{matrix} of
samples from the joint posterior distribution of the regression coefficients
at the sample level
\item \code{theta} an object of class \code{matrix} of
samples from the joint posterior distribution of the sample-level presence
probability, theta. Note that this is only returned if \code{beta_bin} is
\code{TRUE} and \code{sample$model = ~ 1} or \code{sample$model = ~ site}
\item \code{delta} an object of class \code{matrix} of samples from the joint
posterior distribution of the regression coefficients at the replicate level
\item \code{p} an object of class \code{numeric} of sample from the posterior
distribution of the replicate level detection probability. NOte that this is
only returned if \code{beta_bin} is \code{TRUE} and \code{rep$model = ~ 1}.
\item \code{model.info} an object of class list containing the following
elements: \cr
\itemize{
\item \code{X} design matrix for site level predictors
\item \code{W} design matrix for sample level predictors
\item \code{V} design matrix for replicate level predictors
\item \code{M} number of sites
\item \code{J} vector of number of samples per site
\item \code{K} vector of number of replicates per site-sample combination
\item \code{z} matrix of posterior samples of latent site-presence z
\item \code{z.vec} matrix of posterior samples of latent site-presence
stretched across samples
\item \code{A} matrix of posterior samples of latent sample-presence A
\item \code{Y} vector of binomial responses (aggregated at sample level)
\item \code{y} vector of Bernoulli responses (stretched across site-sample
combination)
\item \code{site_mod} model statement for site-level predictors
\item \code{samp_mod} model statement for sample-level predictors
\item \code{rep_mod} model statement for replicate-level predictors
\item \code{num.mcmc} number of MCMC samples run
\item \code{beta_bin} was beta-binomial sampler implemented if possible?
\item \code{df} empty \code{data.frame} of design in long format
}
}
}
\description{
This function fits the Bayesian multi-scale occupancy model
described by \href{https://pubs.er.usgs.gov/publication/70194000}{Dorazio
and Erickson (2017)} using the polya-gamma data augmentation strategy
described by \href{https://arxiv.org/pdf/1205.0310.pdf}{Polson et al.
(2012)}. Note that this documentation assumes there are \eqn{M} sites,
\eqn{J_i} samples within each site, and \eqn{K_{ij}} replicates from each
sample.
}
\details{
This function fits the multi-scale occupancy model described by
\href{https://pubs.er.usgs.gov/publication/70194000}{Dorazio and Erickson
(2017)}. However, this function implements a fully Bayesian sampler based on
the data augmentation strategy described by
\href{https://arxiv.org/pdf/1205.0310.pdf}{Polson et al. (2012)}
}
\examples{
data(fung)

# prep data
fung.detect <- fung \%>\%
  dplyr::select(1:4)

site.df <- fung \%>\%
  dplyr::select(-sample, -pcr1, -pcr2) \%>\%
  dplyr::distinct(site, .keep_all = TRUE) \%>\%
  dplyr::arrange(site)

sample.df <- fung \%>\%
  dplyr::select(-pcr1, -pcr2) \%>\%
  dplyr::arrange(site, sample)

# fit intercept model at all three levels use beta-binomial sampler
fung_mod1 <- msocc_mod(wide_data = fung.detect, progress = T,
                       site = list(model = ~ 1, cov_tbl = site.df),
                       sample = list(model = ~ 1, cov_tbl = sample.df),
                       rep = list(model = ~ 1, cov_tbl = sample.df), # covariates aggregated at sample level
                       num.mcmc = 1000, beta_bin = T)

# model sample level occurence by frog density
fung_mod2 <- msocc_mod(wide_data = fung.detect, progress = T,
                       site = list(model = ~ 1, cov_tbl = site.df),
                       sample = list(model = ~ frogs, cov_tbl = sample.df),
                       rep = list(model = ~ 1, cov_tbl = sample.df),
                       num.mcmc = 1000, beta_bin = T)


}
