% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_mcmc.R
\name{p_mcmc}
\alias{p_mcmc}
\title{Posterior samples of derived rep-level occurence probability}
\usage{
p_mcmc(msocc_mod)
}
\arguments{
\item{msocc_mod}{output from \code{\link{msocc_mod}}}
}
\value{
an object of class \code{matrix} of dimension \code{num.mcmc} by \code{sum(K)}
 of posterior samples of p
}
\description{
This function converts posterior samples from rep-level
 regression coefficients into posterior samples from the derived rep-level
 occurence probability using the logit link function. If p was sampled
 directly using a beta-binomial sampler, if restructures these samples into
 an appropriate format.
}
\details{
This function returns one column for each replicate and arranges the
columns by site, sample, and replicate.
}
\examples{
data(fung)

# prep data
fung.detect <- fung \%>\%
  dplyr::select(1:4)

site.df <- fung \%>\%
  dplyr::select(-sample, -pcr1, -pcr2) \%>\%
  dplyr::distinct(site, .keep_all = TRUE) \%>\%
  dplyr::arrange(site)

sample.df <- fung \%>\%
  dplyr::select(-pcr1, -pcr2) \%>\%
  dplyr::arrange(site, sample)

# fit intercept model at all three levels use beta-binomial sampler
fung_mod1 <- msocc_mod(wide_data = fung.detect, progress = T,
                       site = list(model = ~ 1, cov_tbl = site.df),
                       sample = list(model = ~ 1, cov_tbl = sample.df),
                       rep = list(model = ~ 1, cov_tbl = sample.df), # covariates aggregated at sample level
                       num.mcmc = 1000, beta_bin = T)

p_mcmc(fung_mod1)
}
