% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_summary.R
\name{posterior_summary}
\alias{posterior_summary}
\title{Summarize posterior samples}
\usage{
posterior_summary(MSOcc_mod, burnin = 0, print = T,
  level = "overall", quantiles = c(0.025, 0.975), n = "all")
}
\arguments{
\item{MSOcc_mod}{output from \code{\link{MSOcc_mod}}}

\item{burnin}{number of samples to discard as burnin when summarizing the
posterior}

\item{print}{should summaries be printed in the console?}

\item{level}{one of \code{c('overall', 'site', 'sample', 'rep')}; which level of the
design should be summarized?}

\item{quantiles}{vector of quantiles for credibility intervals}

\item{n}{number of rows to print in the console for the posterior summary}
}
\value{
object of class \code{tbl} providing summaries of the posterior
}
\description{
This function allows for easy summaries of the posterior samples
 drawn by \code{\link{MSOcc_mod}}.
}
\examples{
data(fung)

# prep data
fung_detect <- fung \%>\%
  dplyr::select(1:4)

fung_sitecov_tbl <- fung \%>\%
  dplyr::select(-sample, -pcr1, -pcr2) \%>\%
  unique(.)

fung_sampcov_tbl <- fung \%>\%
  dplyr::select(-pcr1, -pcr2)

# model sample level occurence by frog density
fung_mod2 <- MSOcc_mod(wide_data = fung_detect, progress = T,
                       site = list(model = ~ 1, cov_tbl = fung_sitecov_tbl),
                       sample = list(model = ~ frogs, cov_tbl = fung_sampcov_tbl),
                       num.mcmc = 5000, beta_bin = T)

# overall summary
posterior_summary(fung_mod2)

# print unique rows
posterior_summary(fung_mod2) \%>\%
  dplyr::mutate(site = unlist(lapply(strsplit(site_sample, '_'), function(x) x[1]))) \%>\%
  dplyr::select(-site_sample) \%>\%
  dplyr::select(site, dplyr::everything()) \%>\%
  dplyr::distinct()

# site level
posterior_summary(fung_mod2, level = 'site')

# sample level
posterior_summary(fung_mod2, level = 'sample')

# rep level
posterior_summary(fung_mod2, level = 'rep')

}
