% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psi_mcmc.R
\name{psi_mcmc}
\alias{psi_mcmc}
\title{Posterior samples of derived site-level occurence probability}
\usage{
psi_mcmc(MSOcc_mod)
}
\arguments{
\item{MSOcc_mod}{output from \code{\link{MSOcc_mod}}}
}
\value{
an object of class \code{matrix} of dimension \code{num.mcmc} by \code{M} of
 posterior samples of psi
}
\description{
This function converts posterior samples from site-level
 regression coefficients into posterior samples from the derived site-level
 occurence probability using the logit link function. If psi was sampled
 directly using a beta-binomial sampler, if restructures these samples into
 an appropriate format.
}
\examples{
data(fung)

# prep data
fung_detect <- fung \%>\%
  dplyr::select(1:4)

fung_sitecov_tbl <- fung \%>\%
  dplyr::select(-sample, -pcr1, -pcr2) \%>\%
  unique(.)

fung_sampcov_tbl <- fung \%>\%
  dplyr::select(-pcr1, -pcr2)

# fit intercept model at all three levels use beta-binomial sampler
fung_mod1 <- MSOcc_mod(wide_data = fung_detect, progress = T,
                       site = list(model = ~ 1, cov_tbl = fung_sitecov_tbl),
                       sample = list(model = ~ 1, cov_tbl = fung_sampcov_tbl),
                       num.mcmc = 5000, beta_bin = T)

psi_mcmc(fung_mod1)
}
