% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waic.R
\name{waic}
\alias{waic}
\title{Compute waic}
\usage{
waic(MSOcc_mod, type = 2)
}
\arguments{
\item{MSOcc_mod}{output from \code{\link{MSOcc_mod}}}

\item{type}{one of \code{c(1, 2)} denoting the type of penalty to use when
calculating the waic}
}
\value{
numeric value that is the waic
}
\description{
This function computes the waic described in Gelman et al. (2013)
 for multi-scale occupancy models.
}
\details{
The authors of Gelman et al. (2013) note that the type 2 penalty is a
 better representation of leave one out cross-validation, and therefore
 recommend its use. \cr In the case of hierarchical models, they also note
 that there are two ways to think the likelihood; one that incorporates the
 hyper-parameters and one that does not. Both are arguably justifiable
 depending upon the situation. We do not incorporate the hyper-parameters in
 our calculations here.
}
