% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_lppd.R
\name{compute_lppd}
\alias{compute_lppd}
\title{Compute log pointwise predictive density}
\usage{
compute_lppd(msocc_mod)
}
\arguments{
\item{msocc_mod}{output from \code{\link{msocc_mod}}}
}
\value{
numeric value that is the lppd
}
\description{
This function computes the log pointwise predictive density
 (lppd) described in
 \href{https://link.springer.com/article/10.1007/s11222-013-9416-2}{Gelman et
 al. (2013)} for multi-scale occupancy models.
}
