% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_waic1.R
\name{compute_pwaic1}
\alias{compute_pwaic1}
\title{Compute type 1 waic penalty}
\usage{
compute_pwaic1(MSOcc_mod)
}
\arguments{
\item{MSOcc_mod}{output from \code{\link{MSOcc_mod}}}
}
\value{
numeric value that is the pwaic1
}
\description{
This function computes the type 1 penalty for waic (pwaic1)
 described in Gelman et al. (2013) for multi-scale occupancy models.
}
