% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_waic1.R
\name{compute_pwaic1}
\alias{compute_pwaic1}
\title{Compute type 1 waic penalty}
\usage{
compute_pwaic1(msocc_mod)
}
\arguments{
\item{msocc_mod}{output from \code{\link{msocc_mod}}}
}
\value{
numeric value that is the pwaic1
}
\description{
This function computes the type 1 penalty for waic (pwaic1)
 described in
 \href{https://link.springer.com/article/10.1007/s11222-013-9416-2}{Gelman et
 al. (2013) information criterion paper} for multi-scale occupancy models.
}
