% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_waic2.R
\name{compute_pwaic2}
\alias{compute_pwaic2}
\title{Compute type 2 waic penalty}
\usage{
compute_pwaic2(MSOcc_mod)
}
\arguments{
\item{MSOcc_mod}{output from \code{\link{MSOcc_mod}}}
}
\value{
numeric value that is the pwaic2
}
\description{
This function computes the type 1 penalty for waic (pwaic2)
 described in Gelman et al. (2013) for multi-scale occupancy models.
}
