% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSOcc_mod.R
\name{MSOcc_mod}
\alias{MSOcc_mod}
\title{Fit multi-scale occupancy model.}
\usage{
MSOcc_mod(wide_data, site = list(model = ~1, cov_tbl),
  sample = list(model = ~1, cov_tbl), priors = list(site = list(mu0 =
  0, Lambda0 = 10, eta0 = 0.01, S0 = 1, Sigma0 = 25), sample = list(mu0 =
  0, Lambda0 = 10, eta0 = 0.01, S0 = 1, Sigma0 = 25), rep = list(a0 = 1, b0
  = 1)), num.mcmc = 10000, progress = T, print = NULL, seed = NULL,
  beta_bin = F)
}
\arguments{
\item{wide_data}{object of class \code{data.frame} containing site, sample,
and PCR replicates in wide format. Column names should be \code{site},
\code{sample}, \code{PCR1}, \code{PCR2}, ...}

\item{site}{object of class \code{list} containing the following elements: \cr
\itemize{
\item \code{model} object of class formula describing the within site
model. A hierarchical regression model is also available using \code{lmer}
syntax; currently, only a hierarchy by site is implemented. See details. \cr
\item \code{cov_tbl} object of class \code{data.frame}
containing site specific covariates; \code{cov_tbl} should have \code{M}
rows.
}}

\item{sample}{object of class \code{list} containing the following elements: \cr
\itemize{
\item \code{model} object of class formula describing the within
sample model. A hierarchical regression model is also available using
\code{lmer} syntax; currently, only a hierarchy by site is implemented. See details. \cr
\item \code{cov_tbl} object of class \code{data.frame} containing sample
specific covariates; \code{cov_tbl} should have \code{sum(J)} rows.
}}

\item{priors}{object of class \code{list} containing the following elements: \cr
\itemize{
\item \code{site} \verb{ } object of class \code{list} containing the
following elements: \cr
\itemize{
\item \code{mu0} prior mean for site-level regression coefficients
(if non-hierarchical) or site-level mu (if hierarchical) \cr
\item \code{Lambda0} prior covariance matrix for site-level mu \cr
\item \code{eta0} degrees of freedom for IW prior on Sigma \cr
\item \code{S0} sums of squares for IW prior on Sigma \cr
\item \code{Sigma0} prior covariance matrix for site-level regression coefficients (non-hierarchical)
}
\item \code{sample} \verb{ } object of class \code{list} containing the
following elements: \cr
\itemize{
\item \code{mu0} prior mean for sample-level regression coefficients
(if non-hierarchical) or sample-level mu (if hierarchical) \cr
\item \code{Lambda0} prior covariance matrix for sample-level mu \cr
\item \code{eta0} degrees of freedom for IW prior on Sigma \cr
\item \code{S0} sums of squares for IW prior on Sigma \cr
\item \code{Sigma0} prior covariance matrix for sample-level regression coefficients (non-hierarchical)
}
\item \code{rep} \verb{ } object of class \code{list} containing the
following elements: \cr
\itemize{
\item \code{a0,b0} shape parameters for beta prior on p
}
}}

\item{num.mcmc}{number of MCMC samples}

\item{progress}{should sampling progress be printed?}

\item{print}{interval for printing; defaults to 5 percent of \code{num.mcmc}}

\item{seed}{optional seed for reproducible samples}

\item{beta_bin}{optional; should a beta-binomial sampler be used when possible? This option is considerably faster.}
}
\value{
object of class \code{list} containing the following elements: \cr
\itemize{
\item \code{beta} an object of class \code{matrix} of samples from the joint
posterior distribution of the regression coefficients at the site level
\item \code{psi} an object of class \code{matrix} of samples from the joint
posterior distribution of the site-level presence probability, psi. Note
that this is only returned if \code{beta_bin} is \code{TRUE} and
\code{site$model = ~ 1}
\item \code{alpha} an object of class \code{matrix} of
samples from the joint posterior distribution of the regression coefficients
at the sample level
\item \code{theta} an object of class \code{matrix} of
samples from the joint posterior distribution of the sample-level presence
probability, theta. Note that this is only returned if \code{beta_bin} is
\code{TRUE} and \code{sample$model = ~ 1} or \code{sample$model = ~ site}
\item \code{p} an object of class \code{numeric} of sample from the posterior
distribution of the replicate level detection probability
\item \code{model.info} an object of class list containing the following elements: \cr
\itemize{
\item \code{X} design matrix for site level predictors
\item \code{W} design matrix for sample level predictors
\item \code{M} number of sites
\item \code{J} vector of number of samples per site
\item \code{K} vector of number of replicates per site-sample combination, or numeric value if K is constant
\item \code{z} matrix of posterior samples of latent site-presence z
\item \code{z.vec} matrix of posterior samples of latent site-presence stretched across samples
\item \code{A} matrix of posterior samples of latent sample-presence A
\item \code{Y} observed response matrix
\item \code{site_mod} model statement for site-level predictors
\item \code{samp_mod} model statement for sample-level predictors
\item \code{num.mcmc} number of MCMC samples run
\item \code{beta_bin} was beta-binomial sampler implemented if possible?
}
}
}
\description{
This function fits the Bayesian multi-scale occupancy model
described by Dorazio and Erickson (2017) using the polya-gamma data
augmentation strategy described by Polson et al. (2012). A hierarchical
implementation is also available. Note that this documentation assumes there
are \code{M} sites, \code{J} samples within each site, and \code{K}
replicates from each sample.
}
\details{
This function fits variations of the multi-scale occupancy model
described by Dorazio and Erickson (2017). However, this function implements
a fully Bayesian sampler based on the data augmentation strategy described
by Polson et al. (2012). \cr \cr It also supports hierarchical regression at
the site and sample levels. Currently, only hierarchies defined by
\code{site} are supported. To implement a hierarchical regression model,
define the model using the typical conditional syntax. For example, to fit
an intercept only regression model at the site level with random effects for
each site, define \code{site$model = ~ 1 | site}. Think carefully about your
prior specification, as the regression coefficients are related to
probabilities through the logit link function; small changes in the
magnitude of these coefficients can result in large changes in estimated
probability. \cr \cr Currently, only the site and sample level occurence
probabilities can be modelled by covariates. The replicate level detection
probability is assumed constant.
}
\examples{
data(fung)

# prep data
fung_detect <- fung \%>\%
  dplyr::select(1:4)

fung_sitecov_tbl <- fung \%>\%
  dplyr::select(-sample, -pcr1, -pcr2) \%>\%
  unique(.)

fung_sampcov_tbl <- fung \%>\%
  dplyr::select(-pcr1, -pcr2)

# fit intercept model at all three levels use beta-binomial sampler
fung_mod1 <- MSOcc_mod(wide_data = fung_detect, progress = T,
                       site = list(model = ~ 1, cov_tbl = fung_sitecov_tbl),
                       sample = list(model = ~ 1, cov_tbl = fung_sampcov_tbl),
                       num.mcmc = 5000, beta_bin = T)

# model sample level occurence by frog density
fung_mod2 <- MSOcc_mod(wide_data = fung_detect, progress = T,
                       site = list(model = ~ 1, cov_tbl = fung_sitecov_tbl),
                       sample = list(model = ~ frogs, cov_tbl = fung_sampcov_tbl),
                       num.mcmc = 5000, beta_bin = T)

# intercept model for site level occurence with random effects for site
fung_mod3 <- MSOcc_mod(wide_data = fung_detect, progress = T,
                       site = list(model = ~ 1 | site, cov_tbl = fung_sitecov_tbl),
                       sample = list(model = ~ 1, cov_tbl = fung_sampcov_tbl),
                       num.mcmc = 5000, beta_bin = T)

# intercept model for site level occurence, model sample level occurence
# by frogs with with random effects for site
fung_mod4 <- MSOcc_mod(wide_data = fung_detect, progress = T,
                       site = list(model = ~ 1 | site, cov_tbl = fung_sitecov_tbl),
                       sample = list(model = ~ frogs | site, cov_tbl = fung_sampcov_tbl),
                       num.mcmc = 5000, beta_bin = T)

}
