% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_summary.R
\name{posterior_summary}
\alias{posterior_summary}
\title{Summarize posterior samples}
\usage{
posterior_summary(msocc_mod, burnin = 0, print = F,
  level = "overall", quantiles = c(0.025, 0.975), unique = T)
}
\arguments{
\item{msocc_mod}{output from \code{\link{msocc_mod}} (object of class \code{msocc})}

\item{burnin}{number of samples to discard as burnin when summarizing the
posterior}

\item{print}{should summaries be printed in the console?}

\item{level}{one of \code{c('overall', 'site', 'sample', 'rep')}; which level of the
design should be summarized?}

\item{quantiles}{vector of quantiles for credibility intervals}

\item{unique}{should only unique rows be printed?}

\item{n}{number of rows to print in the console for the posterior summary}
}
\value{
object of class \code{data.frame} providing summaries of the posterior
 organized by site, sample, and replicate.
}
\description{
This function allows for easy summaries of the posterior samples
 drawn by \code{\link{msocc_mod}}.
}
\examples{
data(fung)

# prep data
fung.detect <- fung \%>\%
  dplyr::select(1:4)

site.df <- fung \%>\%
  dplyr::select(-sample, -pcr1, -pcr2) \%>\%
  dplyr::distinct(site, .keep_all = TRUE) \%>\%
  dplyr::arrange(site)

sample.df <- fung \%>\%
  dplyr::select(-pcr1, -pcr2) \%>\%
  dplyr::arrange(site, sample)

# model sample level occurence by frog density
fung_mod2 <- msocc_mod(wide_data = fung.detect, progress = T,
                       site = list(model = ~ 1, cov_tbl = site.df),
                       sample = list(model = ~ frogs, cov_tbl = sample.df),
                       rep = list(model = ~ 1, cov_tbl = sample.df),
                       num.mcmc = 1000, beta_bin = T)

# overall summary
posterior_summary(fung_mod2)

# site level
posterior_summary(fung_mod2, level = 'site')

# sample level
posterior_summary(fung_mod2, level = 'sample')

# rep level
posterior_summary(fung_mod2, level = 'rep')
}
